requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
sentence-transformers
faiss-cpu
PyPDF2
python-dotenv